% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-convert-type.R
\name{convert_type}
\alias{convert_type}
\title{Convert between any 2 supported types}
\usage{
convert_type(src_vec, src_type, dst_type, endian = c("little", "big"),
  first_bit = c("msb", "lsb"), pad = c("right", "left"),
  verbose = FALSE)
}
\arguments{
\item{src_vec}{source values}

\item{src_type}{source type}

\item{dst_type}{destination type}

\item{endian}{'little' or 'big'. default: 'little'}

\item{first_bit}{'msb' or 'lsb'. default: 'msb'}

\item{pad}{which side to add padding bits to make a full byte. 'right'
or 'left'. default: 'right'}

\item{verbose}{default: FALSE}
}
\description{
This converts types directly if available. Otherwise it converts from
one type to another by first converting to the built-in
'raw' type and then converting from 'raw' to the destination
type
}
