% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-swap-endian.R
\name{int32_swap_endian}
\alias{int32_swap_endian}
\title{Swap endianness in a int32 vector}
\usage{
int32_swap_endian(int32_vec)
}
\arguments{
\item{int32_vec}{initial int32 vector}
}
\value{
int32 vector with swapped endianness
}
\description{
Swap endianness in a int32 vector
}
