% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-raw.R, R/01-bit.R, R/08-bit.R, R/16-bit.R,
%   R/24-bit.R, R/32-bit.R
\name{is_raw_vec}
\alias{is_raw_vec}
\alias{is_bit_vec}
\alias{is_logical_vec}
\alias{is_bitstring}
\alias{is_uint8_vec}
\alias{is_int8_vec}
\alias{is_hex8_vec}
\alias{is_chr_vec}
\alias{is_utf8_vec}
\alias{is_uint16_vec}
\alias{is_int16_vec}
\alias{is_hex16_vec}
\alias{is_uint24_vec}
\alias{is_hex24_vec}
\alias{is_int32_vec}
\title{Check if values are of the requested type}
\usage{
is_raw_vec(x)

is_bit_vec(x)

is_logical_vec(x)

is_bitstring(x)

is_uint8_vec(x)

is_int8_vec(x)

is_hex8_vec(x)

is_chr_vec(x)

is_utf8_vec(x)

is_uint16_vec(x)

is_int16_vec(x)

is_hex16_vec(x)

is_uint24_vec(x)

is_hex24_vec(x)

is_int32_vec(x)
}
\arguments{
\item{x}{vector of values to test}
}
\value{
single boolean
}
\description{
Check if values are of the requested type
}
