% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nbytes}
\alias{nbytes}
\title{Get the number of bytes in a vector of the given type}
\usage{
nbytes(vec, type)
}
\arguments{
\item{vec}{vector of values}

\item{type}{type of values}
}
\value{
Number of bytes in the vector
}
\description{
Get the number of bytes in a vector of the given type
}
