% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nelems}
\alias{nelems}
\title{Return the number of elements in a vector of values of the given type}
\usage{
nelems(values, type)
}
\arguments{
\item{values}{values}

\item{type}{type}
}
\value{
length
}
\description{
Return the number of elements in a vector of values of the given type
}
