% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-bit-pad.R
\name{pad_bit_vec}
\alias{pad_bit_vec}
\title{Pad a bit vector so that its length is a multiple of the given value}
\usage{
pad_bit_vec(bit_vec, multiple = 8L, pad = c("left", "right"),
  value = FALSE)
}
\arguments{
\item{bit_vec}{vector of bit values}

\item{multiple}{bit length will be padded to be a multiple of this value. default: 8}

\item{pad}{place padding bits on the left or right. default: left}

\item{value}{value to pad with. TRUE or FALSE. default: FALSE}
}
\value{
padded bit vector
}
\description{
Pad a bit vector so that its length is a multiple of the given value
}
