% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-bit-pad.R
\name{pad_logical_vec}
\alias{pad_logical_vec}
\title{Pad a logical vector so that its length is a multiple of the given value}
\usage{
pad_logical_vec(logical_vec, multiple = 8L, pad = c("left", "right"),
  value = FALSE)
}
\arguments{
\item{logical_vec}{vector of logical values}

\item{multiple}{vector length will be padded to be a multiple of this value. default: 8}

\item{pad}{place padding bits on the left or right. default: left}

\item{value}{value to pad with. TRUE or FALSE. default: FALSE}
}
\value{
padded logical vector
}
\description{
Pad a logical vector so that its length is a multiple of the given value
}
