% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-reverse-bits.R
\name{raw_reverse_bits}
\alias{raw_reverse_bits}
\title{Reverse the bits in each element of a raw vector}
\usage{
raw_reverse_bits(raw_vec)
}
\arguments{
\item{raw_vec}{raw vector}
}
\description{
This pure R version is about 8x slower and uses about 16x as much
memory as the Rcpp version
}
