% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-swap-endian.R
\name{raw_swap_endian}
\alias{raw_swap_endian}
\title{Swap endianness in a raw vector}
\usage{
raw_swap_endian(raw_vec, nbytes)
}
\arguments{
\item{raw_vec}{initial raw vector}

\item{nbytes}{number of bytes to consider for each swap. usually 2, 4, or 8}
}
\value{
raw vector with swapped endianness
}
\description{
Swap endianness in a raw vector
}
