% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{relative_size}
\alias{relative_size}
\title{Relative sizes of types compared to 'raw'}
\format{An object of class \code{numeric} of length 16.}
\usage{
relative_size
}
\description{
Relative sizes of types compared to 'raw'
}
\keyword{datasets}
