% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tidy_bitstring}
\alias{tidy_bitstring}
\title{Tidy bitstring e.g. remove spaces}
\usage{
tidy_bitstring(bitstring)
}
\arguments{
\item{bitstring}{bitstring}
}
\value{
bitstring with spaces removed
}
\description{
Tidy bitstring e.g. remove spaces
}
