% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-swap-endian.R
\name{uint16_swap_endian}
\alias{uint16_swap_endian}
\title{Swap endianness in a uint16 vector}
\usage{
uint16_swap_endian(uint16_vec)
}
\arguments{
\item{uint16_vec}{initial uint16 vector}
}
\value{
uint16 vector with swapped endianness
}
\description{
Swap endianness in a uint16 vector
}
