% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vec_length}
\alias{vec_length}
\title{Get the length of a vector of the given type}
\usage{
vec_length(vec, type)
}
\arguments{
\item{vec}{vector of values}

\item{type}{type of values}
}
\value{
Number of elements in the vector
}
\description{
Get the length of a vector of the given type
}
